requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==4.2.24
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.6.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
Pillow==9.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pypng==0.20220715.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
